requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
defusedxml==0.6.0
Django==3.1.5
django-allauth==0.41.0
django-autoslug==1.9.8
django-cleanup==5.1.0
django-crispy-forms==1.9.1
django-filter==2.4.0
django-phonenumber-field==5.0.0
idna==2.10
itsdangerous==1.1.0
oauthlib==3.1.0
phonenumbers==8.12.13
Pillow==7.1.2
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
